requirements.txt
streamlityfinancepandasnumpyscikit-learn